+++ src/Plugin.php
@@ -54,7 +54,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '3.2.0';
+    public string $schemaVersion = '3.3.0';
 
     /**
      * @inheritdoc